refactor: hooks 폴더 디렉토리 구조 수정
