+++ src/hooks/admin/ApplicationOpenPage/usePeopleAmount.ts
